requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
Pillow
pydantic-settings
PIL